+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.0 (unreleased)
 
 - Updated MIDAS to 1.0.0
+- Added `threshold` option
 - Added `seed` option
 - Changed default `alpha` to 0.5
 - Fixed reading data from files with `directed: false`